tty: Use physical address on Ranchu/Goldfish platforms
Goldfish TTY for Ranchu platforms, has been modified to use
phisical addresses instead of virtual. This change eliminates
the need for snooping through the TLB in QEMU using
cpu_get_phys_page_debug() which does not guarantee that it will
return the valid va -> pa mapping.

This change does not affect the old style Goldfish TTY behaviour
which is still used by the Goldfish emulator. Version register
has been added and probed to see which platform is running this
driver. Reading from the new GOLDFISH_TTY_VERSION register using
the Goldfish emulator it will return 0 and driver will work with
virtual addresses. When run on Ranchu it returns 1 and physical
addresses are used.

Signed-off-by: Miodrag Dinic <miodrag.dinic@imgtec.com>
Signed-off-by: Aleksandar Markovic <aleksandar.markovic@imgtech.com>