MIPS: Add alternative position in memory for appended DTB
Some systems might expect DTB to be located right after the area where
the kernel is loaded in RAM (kernel high address). This commit introduces
configuration option CONFIG_MIPS_DTB_ALTERNATIVE_POSITION, that is
implemented as a suboption of the option MIPS_RAW_APPENDED_DTB, and governs
the choice of the right behaviour related to the DTB placement in memory
described above.

Signed-off-by: Miodrag Dinic <miodrag.dinic@imgtec.com>
Signed-off-by: Aleksandar Markovic <aleksandar.markovic@imgtec.com>