MIPS: Add Ranchu virtual machine for Mips
This commit adds Mips Ranchu virtual machine used by Android
emulator. "Goldfish" and "ranchu" are names for two generations
of virtual boards used by Android emulator. Ranchu is a newer
one among the two, and this commit deals with Ranchu. However,
for historical reasons, some directory, file, device, and variable
names in this commit still contain the word "goldfish". Such name
organization is still present in Android code, and this commit
just reflects it in order to avoid additional differences between
upstream and Android kernels.

Mips Ranchu machine represents a virtual board which is used by
Android emulator. It inherits a number of Goldfish devices which
are modified to be used without the Goldfish platform Bus controller
and are installed using the Simple Bus interface. The support for
Virtio devices is also included. Ranchu board supports up to 16
virtio devices which can be attached using virtio MMIO Bus. This is
summarized in the following picture:

ABUS
 ||----MIPS CPU
 ||      |                      IRQs
 ||----Goldfish PIC-------------------(32)------
 ||                   | | | | | | | |      | |
 ||----Goldfish TTY---- | | | | | | |      | |
 ||                     | | | | | | |      | |
 ||----Goldfish RTC------ | | | | | |      | |
 ||                       | | | | | |      | |
 ||----Goldfish Timer------ | | | | |      | |
 ||                         | | | | |      | |
 ||----Goldfish FB----------- | | | |      | |
 ||                           | | | |      | |
 ||----Goldfish Events--------- | | |      | |
 ||                             | | |      | |
 ||----Goldfish Audio------------ | |      | |
 ||                               | |      | |
 ||----Goldfish Battery------------ |      | |
 ||                                 |      | |
 ||----Android PIPE------------------      | |
 ||                                        | |
 ||----Virtio MMIO Bus                     | |
 ||      |   |   |                         | |
 ||      |   |   (virtio-block)------------- |
 ||    (16)  |                               |
 ||      |   (virtio-net)---------------------

Device tree support is also contained in this commit. Device Tree is
created on the QEMU side based on the information about devices IO map
and IRQ numbers. DTB is subsequently automatically generated and
appended right after the area where the kernel is loaded. Kernel will
pick up this DTB and register everything according to the information
provided. It is assumed that the kernel has following two options emabled:
CONFIG_MIPS_RAW_APPENDED_DTB and CONFIG_MIPS_DTB_ALTERNATIVE_POSITION,
Otherwise, it will use the built-in device tree which may hold stale
information.

Signed-off-by: Miodrag Dinic <miodrag.dinic@imgtec.com>
Signed-off-by: Aleksandar Markovic <aleksandar.markovic@imgtec.com>
Signed-off-by: Goran Ferenc <goran.ferenc@imgtec.com>